# thunder-tests/swagger-collection.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"index","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["health"]}},"/api/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"description":"List of cats","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/check":{"get":{"operationId":"check","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/logout":{"delete":{"operationId":"logout","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/sports":{"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["sports"]},"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSportDto"}}}},"responses":{"201":{"description":""}},"tags":["sports"]}},"/api/sports/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["sports"]},"put":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSportDto"}}}},"responses":{"200":{"description":""}},"tags":["sports"]},"delete":{"operationId":"delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["sports"]}},"/api/sports/{id}/leagues":{"get":{"operationId":"findAllLeagues","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["sports"]}},"/api/leagues":{"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["leagues"]},"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLeagueDto"}}}},"responses":{"201":{"description":""}},"tags":["leagues"]}},"/api/leagues/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["leagues"]},"put":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLeagueDto"}}}},"responses":{"200":{"description":""}},"tags":["leagues"]},"delete":{"operationId":"delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["leagues"]}},"/api/seasons/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["seasons"]}},"/api/seasons":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSeasonDto"}}}},"responses":{"201":{"description":""}},"tags":["seasons"]}},"/api/seasons/{id}/teams":{"get":{"operationId":"findAllTeams","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["seasons"]}},"/api/seasons/{id}/players":{"get":{"operationId":"findAllPlayers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["seasons"]}},"/api/seasons/{id}/matches":{"get":{"operationId":"findAllMatches","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["seasons"]}},"/api/teams/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["teams"]}},"/api/teams":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTeamDto"}}}},"responses":{"201":{"description":""}},"tags":["teams"]}},"/api/teams/{id}/players":{"get":{"operationId":"findAllPlayers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["teams"]}},"/api/players/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["players"]}},"/api/players":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlayerDto"}}}},"responses":{"201":{"description":""}},"tags":["players"]}},"/api/matches/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["matches"]},"put":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMatchDto"}}}},"responses":{"200":{"description":""}},"tags":["matches"]},"delete":{"operationId":"delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["matches"]}},"/api/matches":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMatchDto"}}}},"responses":{"201":{"description":""}},"tags":["matches"]}},"/api/matches/{id}/players":{"get":{"operationId":"findAllPlayers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["matches"]}},"/api/users":{"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]},"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/api/users/{id}":{"get":{"operationId":"findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]},"put":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]}}},"info":{"title":"iSports","description":"The iSports API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"tyler.beutel@blackbook.ai"},"password":{"type":"string","example":"Aa@123456"}},"required":["email","password"]},"CreateSportDto":{"type":"object","properties":{}},"UpdateSportDto":{"type":"object","properties":{}},"CreateLeagueDto":{"type":"object","properties":{}},"UpdateLeagueDto":{"type":"object","properties":{}},"CreateSeasonDto":{"type":"object","properties":{}},"CreateTeamDto":{"type":"object","properties":{}},"CreatePlayerDto":{"type":"object","properties":{}},"CreateMatchDto":{"type":"object","properties":{}},"UpdateMatchDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}}}}}